240809 (오후) 별점 toggle 끝

+++ src/main/java/com/movie/star/BO/StarBO.java
@@ -23,10 +23,15 @@ public class StarBO {
 	// 별점 toggle 
 	public void starToggle(int movieId, int userOriginId, int point) {
 		int count = starMapper.selectStarCountByMovieIdAndUserOriginId(movieId, userOriginId);
-		if (count == 0) { // 아직 별점을 안누른 상태 > 클릭 > 별점 매김
+		Integer pastPoint = (Integer) starMapper.selectPointByMovieIdAndUserOriginId(movieId, userOriginId);
+		if (count == 0) { // 아직 별점을 안누른 상태
 			starMapper.insertStarByMovieIdAndUserOriginIdAndPoint(movieId, userOriginId, point);
-		} else { // 이미 별점 누른 상태 > 재클릭 > 별점 삭제 
-			starMapper.deleteStarByMovieIdAndUserOriginId(movieId, userOriginId);
+		} else { // 이미 별점 누른 상태
+			if (point == pastPoint) { // 내가 방금 누른 n번별(point)과 이미 매겨져있는 점수(pastPoint)가 동일한 경우 (즉, 재클릭)
+				starMapper.deleteStarByMovieIdAndUserOriginId(movieId, userOriginId); // 재클릭 > 별점 삭제
+			} else { // 내가 방금 누른 n번별(point)과 이미 매겨져있는 점수(pastPoint)가 다른 경우 (즉, 점수 수정 = 높이거나 낮추거나)
+				starMapper.updateStarByMovieIdAndUserOriginIdAndPoint(movieId, userOriginId, point); // 별점 수정
+			}
 		}
 	}
 	
@@ -35,13 +40,23 @@ public class StarBO {
 		return starMapper.insertStarByMovieIdAndUserOriginIdAndPoint(movieId, userOriginId, point);
 	}
 	
+	// 별점 수정하기 - db에 update
+	public int updateStarByMovieIdAndUserOriginIdAndPoint(int movieId, int userOriginId, int point) {
+		return starMapper.updateStarByMovieIdAndUserOriginIdAndPoint(movieId, userOriginId, point);
+	}
+	
 	// 별점 삭제하기 - db에서 delete
 	public int removeStarByMovieIdAndUserOriginId(int movieId, int userOriginId) {
 		return starMapper.deleteStarByMovieIdAndUserOriginId(movieId, userOriginId);
 	}
 	
+	// 특정 유저가 특정 영화에 '몇 점'을 매겼나 알아내는 메소드 - db에서 int(point) select
+	public int getPointByMovieIdAndUserOriginId(int movieId, int userOriginId) {
+		return starMapper.selectPointByMovieIdAndUserOriginId(movieId, userOriginId);
+	}
+	
 	// 특정 영화의 특정 유저가 매긴 별점 정보 일단 가져오기 - db에서 select
-	public List<Star> getStarInfoByMovieIdAndUserOriginId(int movieId, int userOriginId) {
+	public Star getStarInfoByMovieIdAndUserOriginId(int movieId, int userOriginId) {
 		return starMapper.selectStarInfoByMovieIdAndUserOriginId(movieId, userOriginId);
 	}
